Fix several edit use bugs
